navbar moved under components and round buttons
